chore: remove unnecassary files and rename package as webconsole
