--- webconsole/api_default.go
@@ -1,57 +1,0 @@
-// SPDX-FileCopyrightText: 2025 Canonical Ltd
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-/*
- * WebConsole NFConfig API
- *
- * API for managing access, mobility, policy, session and PLMN information.
- *
- * API version: 1.0.0
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-import (
-	"github.com/gin-gonic/gin"
-)
-
-type DefaultAPI struct {
-}
-
-// Get /nfconfig/access-mobility
-// Get Access and Mobility Configuration
-func (api *DefaultAPI) NfconfigAccessMobilityGet(c *gin.Context) {
-	// Your handler implementation
-	c.JSON(200, gin.H{"status": "OK"})
-}
-
-// Get /nfconfig/plmn
-// Get PLMN ID Configuration
-func (api *DefaultAPI) NfconfigPlmnGet(c *gin.Context) {
-	// Your handler implementation
-	c.JSON(200, gin.H{"status": "OK"})
-}
-
-// Get /nfconfig/plmn-snssai
-// Get PLMN-SNSSAI Configuration
-func (api *DefaultAPI) NfconfigPlmnSnssaiGet(c *gin.Context) {
-	// Your handler implementation
-	c.JSON(200, gin.H{"status": "OK"})
-}
-
-// Get /nfconfig/policy-control
-// Get Policy Control Configuration
-func (api *DefaultAPI) NfconfigPolicyControlGet(c *gin.Context) {
-	// Your handler implementation
-	c.JSON(200, gin.H{"status": "OK"})
-}
-
-// Get /nfconfig/session-management
-// Get Session Management Configuration
-func (api *DefaultAPI) NfconfigSessionManagementGet(c *gin.Context) {
-	// Your handler implementation
-	c.JSON(200, gin.H{"status": "OK"})
-}
