+++ webconsole/model_access_and_mobility.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type AccessAndMobility struct {
 	PlmnId PlmnId `json:"plmnId,omitempty"`
