+++ webconsole/model_arp.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type Arp struct {
 	PriorityLevel int32 `json:"priorityLevel,omitempty"`
