+++ webconsole/model_dnn_qos.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type DnnQos struct {
 	DnnName string `json:"dnnName,omitempty"`
