+++ webconsole/model_ip_domain.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type IpDomain struct {
 	DnnName string `json:"dnnName,omitempty"`
