+++ webconsole/model_pcc_flow.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type PccFlow struct {
 	Description string `json:"description,omitempty"`
