+++ webconsole/model_pcc_qos.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type PccQos struct {
 	FiveQi int32 `json:"fiveQi,omitempty"`
