+++ webconsole/model_pcc_rule.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type PccRule struct {
 	RuleId string `json:"ruleId,omitempty"`
