+++ webconsole/model_plmn_id.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type PlmnId struct {
 	Mcc string `json:"mcc,omitempty"`
