+++ webconsole/model_plmn_snssai.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type PlmnSnssai struct {
 	PlmnId PlmnId `json:"plmnId,omitempty"`
