+++ webconsole/model_policy_control.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type PolicyControl struct {
 	PlmnId PlmnId `json:"plmnId,omitempty"`
