+++ webconsole/model_session_management.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type SessionManagement struct {
 	SliceName string `json:"sliceName,omitempty"`
