+++ webconsole/model_snssai.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type Snssai struct {
 	Sst int32 `json:"sst,omitempty"`
