+++ webconsole/model_upf.go
@@ -12,7 +12,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package openapi
+package webconsole
 
 type Upf struct {
 	Hostname string `json:"hostname,omitempty"`
